Changes in rest and user controller
Set base rest controller.
some changes in user model for rest.
delete not used controllers.
delete readme and license files from yii2.
