+++ models/User.php
@@ -3,9 +3,14 @@
 namespace app\models;
 
 use Yii;
+use yii\web\IdentityInterface;
+use yii\db\ActiveRecord;
 
+use yii\web\Link;
+use yii\web\Linkable;
+use yii\helpers\Url;
 
-class User extends \yii\db\ActiveRecord implements \yii\web\IdentityInterface
+class User extends ActiveRecord implements IdentityInterface, Linkable
 {
     /**
      * @inheritdoc
@@ -108,4 +113,22 @@ class User extends \yii\db\ActiveRecord implements \yii\web\IdentityInterface
     public function isAuthor($id){
            return $id===Yii::$app->user->id ? true: false;
     }
+	
+	public function getLinks()
+    {
+        return [
+            Link::REL_SELF => Url::to(['user/view', 'id' => $this->id], true),
+        ];
+    }
+	
+	// отбрасываем некоторые поля. Сделано для рест апи, не ясно на что еще влияет
+	public function fields()
+	{
+		$fields = parent::fields();
+
+		// удаляем не безопасные поля
+		unset($fields['password'], $fields['auth_key'], $fields['access_token']);
+
+		return $fields;
+	}
 }
